Get all jobs

+++ job-scheduler/job-service/internal/adapter/http/middleware/auth_middleware.go
@@ -43,7 +43,7 @@ func NewAuthMiddleware(
 			claims := jwt.ExtractClaims(c)
 
 			return &entity.User{
-				Id: claims[userIDKey].(uint64),
+				Id: uint64(claims[userIDKey].(float64)),
 			}
 		},
 		Authenticator: func(c *gin.Context) (interface{}, error) {
